backend/pyproject.toml
[project]
name = "golden-compass-backend"
version = "0.1.0"
description = "Golden Compass Content Generation System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "pydantic>=2.5.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "celery[redis]>=5.3.4",
    "redis>=5.0.1",
    "openai>=1.8.0",
    "elevenlabs>=0.2.27",
    "moviepy>=1.0.3",
    "pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]


[tool.uv]
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]
